+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -296,6 +296,7 @@ Provider = Literal[
     "elevenlabs",
     "litellm",
     "openailike",
+    "cerebras",
 ]
 
 CreatePromptRole = Literal[
@@ -960,6 +961,7 @@ CreatePromptProvider = Literal[
     "elevenlabs",
     "litellm",
     "openailike",
+    "cerebras",
 ]
 
 CreatePromptPromptsRole = Literal[