+++ src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0BTG04M8FDGM97ASKQ124Z1"
+        "01K0RJ8MDG1ZGGZ35WDC47NKC4"
     )
     r"""The id of the resource"""