+++ src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0BTG04N5CBAZ8T5551SWZ14"
+        "01K0RJ8MDHKSQ1YYZEBRV6A4C5"
     )
     r"""The id of the resource"""